requirements.txt
# Core dependencies
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0

# AI/ML models
openai>=1.0.0
transformers>=4.35.0
torch>=2.0.0
accelerate>=0.25.0
sentencepiece>=0.1.99
protobuf>=3.20.0

# Face recognition
face-recognition>=1.3.0
dlib>=19.24.0

# Langfuse tracking
langfuse>=2.0.0

# Async support
aiohttp>=3.9.0

